requirements.txt
neuraloperator>=1.0.0
numpy>=1.25,<2.0
torch>=2.4.0,<2.5.0

matplotlib==3.6.2
scikit-learn==1.1.3
pandas==2.0.1

setuptools>=65.5.0
sympy==1.11.1
tqdm==4.66.2
seaborn>=0.13
pyyaml>=6.0